+++ src/main/java/com/anesuandtatenda/studentlecturerplatform/web/requests/AccountRequest.java
@@ -9,7 +9,7 @@ public class AccountRequest {
     private String firstName;
     private String lastName;
     private String regNumber;
-    private int year;
+    private long year;
     private long programId;
     private Role role;
     private String password;
